affirmations user fix

+++ src/pages/AffirmationsPage/AffirmationsPage.jsx
@@ -5,18 +5,16 @@ import AffirmationCard from "../../components/AffirmationCard/AffirmationCard";
 import * as affirmationAPI from "../../services/affirmations-api";
 import EditAffirmationCard from "../../components/EditAffirmationCard/EditAffirmationCard";
 import AddAffirmationCard from "../../components/AddAffirmationCard/AddAffirmationCard";
-import Quotes from '../../components/Quotes/Quotes'
+import Quotes from "../../components/Quotes/Quotes";
 import { useStateWithCallback } from "../../hooks/useStateWithCallback";
 
 // This is our Affirmation list page! All our functions will live here, and we'll pass from props to components. Here we import all the things we are exporting from all our pages
-function AffirmationPage(props) {
+function AffirmationPage({ user }) {
   // Creating state for Affirmations
+  const [person, setPerson] = useState(user);
   const [affirmations, setAffirmations] = useState([]);
   const history = useHistory();
-  useEffect(() => {
-    // This is listening for changes in affirmations state, then the function below will reroute
-    history.push("/affirmationpage");
-  }, [affirmations, history]);
+
   // Add an affirmation
   async function handleAddAffirmation(newAffirmationData) {
     const newAffirmation = await affirmationAPI.create(newAffirmationData);
@@ -24,7 +22,9 @@ function AffirmationPage(props) {
   }
   // Update an affirmation
   async function handleUpdateAffirmation(updatedAffirmationData) {
-    const updatedAffirmation = await affirmationAPI.update(updatedAffirmationData);
+    const updatedAffirmation = await affirmationAPI.update(
+      updatedAffirmationData
+    );
     const newAffirmationsArray = affirmations.map((m) =>
       m._id === updatedAffirmation._id ? updatedAffirmation : m
     );
@@ -32,14 +32,22 @@ function AffirmationPage(props) {
   }
   // Delete a Affirmation
   async function handleDeleteAffirmation(id) {
-    await affirmationAPI.deleteOne(id);
-    setAffirmations(affirmations.filter((a) => a._id !== id));
+    if (user) {
+      await affirmationAPI.deleteOne(id);
+      setAffirmations(affirmations.filter((a) => a._id !== id));
+    } else {
+      history.push("/login");
+    }
   }
   /*--- Lifecycle Methods ---*/
   useEffect(() => {
     (async function () {
-      const affirmations = await affirmationAPI.getAll();
-      setAffirmations(affirmations);
+      const allAffirmations = await affirmationAPI.getAll();
+      setAffirmations(
+        allAffirmations.filter(
+          (affirmation) => user._id === affirmation.postedBy._id
+        )
+      );
     })();
   }, []);
 
@@ -53,33 +61,35 @@ function AffirmationPage(props) {
       </>
 
       <>
-      <Quotes />
+        <Quotes />
       </>
 
       <div>
-      <>
-      <>
-        <AddAffirmationCard
-        affirmation={affirmations.length}
-         handleAddAffirmation={handleAddAffirmation}
-         />
-        </>
+        <>
 
-        {affirmations.map(affirmation => (
-         <p>
-         <AffirmationCard
-         affirmation={affirmation}
-         handleDeleteAffirmation={handleDeleteAffirmation}
-         key={affirmation._id} />
-          <EditAffirmationCard
-            affirmation={affirmation}
-            handleUpdateAffirmation={handleUpdateAffirmation}
-            key={affirmation._id}
-          />
-         </p>
-        ))}
-        </>
+            <AddAffirmationCard
+              user={person}
+              affirmation={affirmations.length}
+              handleAddAffirmation={handleAddAffirmation}
+            />
 
+          {affirmations.length ? (
+          <>
+          {affirmations.map((affirmation) => (
+            <p>
+              <AffirmationCard
+                affirmation={affirmation}
+                handleDeleteAffirmation={handleDeleteAffirmation}
+                handleUpdateAffirmation={handleUpdateAffirmation}
+                key={affirmation._id}
+              />
+            </p>
+          ))}
+          </>
+          ) : (
+            <p>Add some affirmations</p>
+          )}
+        </>
       </div>
     </div>
   );
